+++ src/components/MatchCriterion/edit/IncomingSectionStatic/index.js
@@ -1,0 +1,5 @@
+export * from './IncomingSectionStatic';
+export * from './StaticValueDate';
+export * from './StaticValueDateRange';
+export * from './StaticValueNumber';
+export * from './StaticValueText';