+++ src/components/MatchCriterion/view/IncomingSectionStatic/StaticValueNumber.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { KeyValue } from '@folio/stripes/components';
+
+export const StaticValueNumber = ({ value }) => {
+  return (
+    <div data-test-static-number-field>
+      <KeyValue value={value} />
+    </div>
+  );
+};